// chifa-site/src/payload-types.ts
export interface Media {
  id: string;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * Colección de platos del menú del restaurante.
 *
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "platos".
 */
export interface Plato {
  id: string;
  nombre: string;
  descripcion: string;
  precio: number;
  categoria: string | Categoria;
  imagen?: (string | null) | Media;
  recomendado?: boolean | null;
  picante?: ('no_picante' | 'suave' | 'medio' | 'picante' | 'muy_picante') | null;
  /**
   * Listar alérgenos separados por comas.
   */
  alergenos?: string | null;
  etiquetas_dieteticas?: ('vegetariano' | 'vegano' | 'sin_gluten')[] | null;
  tipo_proteina?: ('pollo' | 'carne_de_res' | 'cerdo' | 'pescado' | 'mariscos' | 'tofu' | 'otro' | 'ninguno') | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * Categorías para agrupar los platos del menú.
 *
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "categorias".
 */
export interface Categoria {
  id: string;
  nombre: string;
  updatedAt: string;
  createdAt: string;
}